requirements.runtime.txt
# Runtime-only dependencies for API inference container
# Generated from requirements.txt by keeping only production inference/API packages.

annotated-types==0.7.0
anyio==4.11.0
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
confusable-homoglyphs==3.3.1
faiss-cpu==1.13.2
fastapi==0.120.1
filelock==3.20.3
ftfy==6.3.1
fsspec==2025.10.0
google-cloud-storage>=2.10.0
gcsfs>=2024.1.0
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
joblib==1.5.2
numpy==2.3.4
packaging==25.0
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.6
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.3.0
sniffio==1.3.1
starlette==0.50.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.38.0
xgboost==3.1.1
